Fix cppcheck and fix cppcheck errors
cppcheck regularly changes its interface and this time it changed the
checks that are run per default, so no meaningful checks were run.

* Reenable the relevant cppchecks
* Fix the errors now returned by cppcheck
* Following types of errors are fixed:
  * scope: variables can be moved to a inner scope
  * printf format strings: incorrect formats used
  * casts: quickcheck uses casts to and from void*, but the buffer is
    defined as char* to allow pointer arithmetic. So we have to cast to
    void* AFTER pointer arithmetic is done.

Note: we have to replace quickcheck with something generic, I hate all
this casting.

+++ src/chirp.c
@@ -540,15 +540,15 @@ _ch_chirp_verify_cfg(ch_chirp_t* chirp)
     V(
         chirp,
         conf->BUFFER_SIZE >= CH_MIN_BUFFER_SIZE || conf->BUFFER_SIZE == 0,
-        "Config: buffer size must be > %d (%d)",
+        "Config: buffer size must be > %d (%u)",
         CH_MIN_BUFFER_SIZE,
         conf->BUFFER_SIZE
     );
     V(
         chirp,
         conf->BUFFER_SIZE >= sizeof(ch_message_t) || conf->BUFFER_SIZE == 0,
-        "Config: buffer size must be > %d (%d)",
-        (int) sizeof(ch_message_t),
+        "Config: buffer size must be > %lu (%u)",
+        (unsigned long) sizeof(ch_message_t),
         conf->BUFFER_SIZE
     );
     V(
@@ -556,8 +556,8 @@ _ch_chirp_verify_cfg(ch_chirp_t* chirp)
         conf->BUFFER_SIZE >= sizeof(
             ch_rd_handshake_t
         ) || conf->BUFFER_SIZE == 0,
-        "Config: buffer size must be > %d (%d)",
-        (int) sizeof(ch_rd_handshake_t),
+        "Config: buffer size must be > %lu (%u)",
+        (unsigned long) sizeof(ch_rd_handshake_t),
         conf->BUFFER_SIZE
     );
     return CH_SUCCESS;
